scripts/streamlit_app/requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.2
Pillow>=10.0.0
python-dotenv>=1.0.1
